ermmmm what the sigma

+++ backend/controllers/cartController.js
@@ -21,16 +21,31 @@ const removeFromCart = async (req, res) => {
     try {
         const { user_id, coffee_id } = req.body;
         await pool.query(
-            `UPDATE cart_items
-             SET quantity = quantity - 1
-             WHERE user_id = $1 AND coffee_id = $2 AND quantity > 0`,
+            `UPDATE cart_items SET quantity = quantity - 1 WHERE user_id = $1 AND coffee_id = $2 AND quantity > 0`,
             [user_id, coffee_id]
         );
         // Deletes the item from the cart if quantity is equal or less than zero
         await pool.query(
-            `DELETE FROM cart_items
-             WHERE user_id = $1 AND coffee_id = $2 AND quantity <= 0`,
+            `DELETE FROM cart_items WHERE user_id = $1 AND coffee_id = $2 AND quantity <= 0`,
             [user_id, coffee_id]
+        );
+        res.json({ success: true, message: 'Removed From Cart' });
+    } catch (error) {
+        console.error(error);
+        res.json({ success: false, message: 'Error' });
+    }
+};
+
+const removeCartById = async (req, res) => {
+    try {
+        const {user_id, cart_id} = req.body;
+        await pool.query(
+            'UPDATE cart_items SET quantity = cart_items.quantity - 1 WHERE user_id = $1 AND cart_id = $2 AND quantity > 0',
+            [user_id, cart_id]
+        );
+        await pool.query(
+            `DELETE FROM cart_items WHERE user_id = $1 AND cart_id = $2 AND quantity <= 0`,
+            [user_id, cart_id]
         );
         res.json({ success: true, message: 'Removed From Cart' });
     } catch (error) {
@@ -58,5 +73,6 @@ const getCart = async (req, res) => {
 module.exports = {
     addToCart,
     removeFromCart,
+    removeCartById,
     getCart
 }
